[FIX] web: Fix actions for widgets w/o controllers
Execution of actions is usually done with trigger_up, which delegates
the action execution to the widget's controller.

This is troublesome for actions done on widgets that have no
controllers, like widgets in Studio for example.

The commit fixes this by adding a standalone option for AbstractFields,
if the flag is true when instantiating a widget, the action will be
called directly from within the widget object and will not be propagated
upwards.

A test to verify that this works properly has been added as well