[9.0][FIX] Correct check of property 'rules'
Current behavior before PR:
In some browsers, like Chrome 64, there is an exception when the property 'rules' is not located in CSS sheets and the user tries to enter to some sections, like 'Preferences'.

Desired behavior after PR is merged:
'rules' property is correctly checked and the execution continues in every browsers.