[FIX] Correct check of property 'rules'
Description of the issue/feature this PR addresses:
The property 'rules' is not located in CSS sheets.

Current behavior before PR:
In some browsers, like Chrome 64, there is an exception when the property 'rules' is not located in CSS sheets and the user tries to enter to some sections, like 'Preferences'.

Desired behavior after PR is merged:
'rules' property is correctly checked and the execution continues in every browsers.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr